+++ src/jwt_auth.py
@@ -7,4 +7,4 @@ def generate_token(traveler, key):
     return jwt.encode({"traveler":traveler, "exp":datetime.datetime.utcnow() + datetime.timedelta(minutes=60)},key)
 
 def decode_token():
-    pass
+    return jwt.decode(token, key, algorithms="HS256")
