+++ src/main.py
@@ -9,16 +9,18 @@ from flask_cors import CORS
 from utils import APIException, generate_sitemap
 from admin import setup_admin
 from models import db, Traveler, Trip, Userpro, Offers
-from encrypted import encrypted_pass, compare_pass
 
 #from models import Person
 
 from werkzeug.utils import secure_filename
 from werkzeug.datastructures import ImmutableMultiDict
 
-from jwt_auth import generate_token, decode_token
+from router.professional import professional_route
+from router.traveler import traveler_route
+from router.trips import trips_route
+import jwt
 
-from functools import wraps
+from functools import wraps #importacion para generar el decorador
 
 app = Flask(__name__)
 app.url_map.strict_slashes = False
@@ -30,21 +32,32 @@ db.init_app(app)
 CORS(app)
 setup_admin(app)
 
-usuario_fake = {
-    "email": "mariocepedaortega@gmail.com",
-    "password": encrypted_pass("123456")
-}
 #decorador
-#def token_required(f):
-    #@wraps(f)
-    #def decorador(*args, **kwargs):
-        #try:
-            #auth = request.headers.get('Authorization')
-            #print(auth)
-        #except:
-            #pass
-        #return f(*args, **kwargs)
-    #return decorador
+def token_required(f):
+    @wraps(f)
+    def decorador(*args, **kwargs):
+        try:
+            auth = request.headers.get('Authorization')
+            print(auth)
+            if auth is None:
+                return jsonify('no token'),403
+            token = auth.split(' ')
+            print(token)
+            data = decode_token(token[1], app.config['SECRET_KEY'])
+            print(data,'@@@@@@@@@@@@@@')
+            traveler = Traveler.query.get(data['traveler']['id_traveler'])#Una vez validado el token compruebo que el traveler realmente pertenece a mi bbd
+            print(traveler)
+            if traveler is None:
+                return jsonify("no authorization"), 401
+        except OSError as error:
+            print(error)
+
+        except jwt.exception.ExpiredSignatureError as err:
+            print(err)
+            return jsonify("token expired"), 403
+            
+        return f(*args, **kwargs)
+    return decorador
 
 # Handle/serialize errors like a JSON object
 @app.errorhandler(APIException)
@@ -56,81 +69,10 @@ def handle_invalid_usage(error):
 def sitemap():
     return generate_sitemap(app)
 
-
-@app.route('/user/register/pro', methods=['POST'])
-def handle_pro():
-    body = request.get_json()
-    new_user = Userpro(user_name=body['user_name'], email=body['email'],password=body['password'],phone=body['phone'],url=body['url'],location=body['location'],direction=body['direction'],vat_number=body['vat_number'],social_reason=body['social_reason'])
-    db.session.add(new_user) #sin este linea no se añade a la base de datos
-    db.session.commit() # esta es la hermana de la que esta arriba :) 
-    print(new_user.serialize())
-    return jsonify(new_user.serialize()),200
-
-@app.route('/user/pro', methods=['GET'])
-def handle_hello():
-    body = Userpro.query.all()
-    for x in body:
-        print(x.serialize())
-    print(body)
-
-@app.route('/user/traveler', methods=['POST'])
-def handle_Traveler():
-    body= request.get_json()
-    new_user = Traveler(username=body["username"],email=body["email"],password=body["password"])
-    db.session.add(new_user)
-    db.session.commit()
-    print(new_user.serialize())
-
-    return jsonify("todo bien"), 200
-    
-@app.route('/pro/<int:id>',methods=['GET'])
-def get_pro(id):
-    user_pro_by_id = Userpro.query.filter_by(id=id).filter_by(is_active=False).first()
-    print(user_pro_by_id)
-    if user_pro_by_id is not None:
-        return jsonify(user_pro_by_id.serialize()), 200
-    else:
-        return jsonify("usuario no existe"),400
-
-@app.route('/viajes', methods=['GET'])
-def get_viajes():
-    total_viajes = Trip.query.all()
-
-    response_body = {
-        "msg": "estos son todos los viajes"
-    }
-
-    return jsonify(response_body), 200
-
-@app.route('/viaje', methods=['POST'])
-def post_trip():
-    body = request.get_json()
-    print(Trip(needs_trip=body['needs_trip'], destination=body['destination'], first_day=body['first_day'], last_day=body['last_day'], description=body['description']))
-    new_trip = Trip(needs_trip=body['needs_trip'], destination=body['destination'], first_day=body['first_day'], last_day=body['last_day'], description=body['description'])
-    db.session.add(new_trip)
-    db.session.commit()
-    print(new_trip.serialize())
-
-    return jsonify(new_trip.serialize()), 200
-
-@app.route('/traveler/login', methods=['POST'])
-def login_traveler():
-    body = request.get_json()
-    print(usuario_fake)
-    #traveler = Traveler.query.filter_by(email=body['email']).first()
-    if(usuario_fake["email"]==body["email"]):
-        traveler = usuario_fake
-    else:
-        traveler = None 
-    if(traveler is None):
-        return "el usuario no existe", 401
-    is_validate = compare_pass(body['password'], traveler["password"].decode("utf-8"))
-    if(is_validate == False):
-       return "password incorrecto", 401
-
-    token = generate_token(traveler["email"], app.config['SECRET_KEY'])
-    #print(token)
-    return jsonify({"access_token":token}), 200
+#llamo a las funciones creadas en la carpeta route y las paso los parámetros app y token_required
+proffesional = professional_route(app,token_required)
+traveler = traveler_route(app,token_required)
+trips = trips_route(app,token_required)
 
 # this only runs if `$ python src/main.py` is executed
 if __name__ == '__main__':
